tugas project

+++ adminpanel/kategori-detail.php
@@ -76,6 +76,18 @@
                 }
                 
                 if(isset($_POST['deleteBtn'])){
+                    $queryCheck = mysqli_query($mysqli, "SELECT * FROM produk WHERE kategori_id='$id'");
+                    $dataCount = mysqli_num_rows($queryCheck);
+                    
+                    if($dataCount>0){
+                        ?>
+                            <div class="alert alert-warning mt-3" role="alert">
+                                Kategori tidak bisa dihapus karena sudah digunakan di produk
+                            </div>
+                        <?php 
+                        die();
+                    }
+
                     $queryDelete = mysqli_query($mysqli, "DELETE FROM kategori WHERE id='$id'");
                     
                     if($queryDelete){
